Basic funktionality
to be used from the command line, with 2 or more arguments.
The first argument is always a port name (e.g. "COM5"), the second and further arguments are the command like "~0000 1".
Each command will be automatically terminated with a carriage return symbol. The second and further commands will be send with a 3 sec delay.
Example: OptoCOM.py "COM2" "~0000 1" "~00141 1" "~00112 3"